# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the core API for Sispeg Project in Azmeela",
        "title": "Sispeg API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reports/monthly-summary": {
            "get": {
                "description": "Get a summary of attendances, overtimes, leaves, and debts for a specific month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get monthly summary report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Month (1-12), defaults to current",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Year (YYYY), defaults to current",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_azmeela_sispeg-api_internal_domain.MonthlySummaryReport"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_azmeela_sispeg-api_internal_domain.MonthlySummaryReport": {
            "type": "object",
            "properties": {
                "total_attendances": {
                    "type": "integer"
                },
                "total_debts": {
                    "type": "integer"
                },
                "total_leaves": {
                    "type": "integer"
                },
                "total_overtimes": {
                    "type": "integer"
                },
                "total_sick_days": {
                    "type": "integer"
                }
            }
        },
...
    }
}